backend/gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
httpx==0.25.2
celery==5.3.4
asyncpg==0.29.0
pydantic[email]==2.5.0
python-dotenv==1.0.0